Add files via upload
This is the project for HID global by Anthony Gutilla and Stephen Harding. Open the workstationform_v2.html to view the page. The initial values are loaded in from an XML file hosted on AWS. Because we did not have access to an actual database, the page allows you to download the edited XML file. You can use the choose file button to upload an XML file and the page will load the data to be edited. Thanks to HID global, Andrea Nigels, Marco Rodriguez, and Dr. Virginia for giving us this fantastic real world project.